Unify attach_indices
