+++ src/details/ArborX_PairValueIndex.hpp
@@ -31,55 +31,6 @@ struct PairValueIndex
   Index index;
 };
 
-namespace Experimental
-{
-template <typename Values, typename Index>
-class AttachIndices
-{
-private:
-  using Data = Details::AccessValues<Values, PrimitivesTag>;
-
-public:
-  Data _data;
-
-  using memory_space = typename Data::memory_space;
-  using value_type = PairValueIndex<typename Data::value_type, Index>;
-
-  AttachIndices(Values const &values)
-      : _data{values}
-  {}
-
-  KOKKOS_FUNCTION
-  auto operator()(int i) const { return value_type{_data(i), Index(i)}; }
-
-  KOKKOS_FUNCTION
-  auto size() const { return _data.size(); }
-};
-
-// Make sure the default Index matches the default in PairValueIndex
-template <typename Index = typename PairValueIndex<int>::index_type,
-          typename Values = void>
-auto attach_indices(Values const &values)
-{
-  return AttachIndices<Values, Index>{values};
-}
-} // namespace Experimental
-
 } // namespace ArborX
-
-template <typename Values, typename Index>
-struct ArborX::AccessTraits<ArborX::Experimental::AttachIndices<Values, Index>,
-                            ArborX::PrimitivesTag>
-{
-  using Self = ArborX::Experimental::AttachIndices<Values, Index>;
-
-  using memory_space = typename Self::memory_space;
-
-  KOKKOS_FUNCTION static auto size(Self const &values) { return values.size(); }
-  KOKKOS_FUNCTION static decltype(auto) get(Self const &values, int i)
-  {
-    return values(i);
-  }
-};
 
 #endif
